GOCBC-1617: Retry on reading first row for query and search
Motivation
----------
The initial call to query or search may succeed but, due to how grpc
works, errors like "service not available" might only be surfaced
when reading the first item from the stream. We need to be able
to retry when this happens.

Changes
-------
Update our couchbase2 request path to accept a "peek" callback
which can be used to access the result before leaving the retry
logic, and respond with an error if required.

Change-Id: I447e0f99887a2c0889d84930cc95de76121adfd6
Reviewed-on: https://review.couchbase.org/c/gocb/+/207633
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>