GOCBC-1626: Handle operations after close gracefully
Motivation
----------
At the moment if cluster close is called twice or operations are performed
after close then the SDK can panic.

Changes
-------
Add logic to connection managers to fast fail any ops when a provider
is fetched and the manager is closed.
Add logic to return error when connection managers are closed twice.
Add provider controllers which are used with helper functions
which increment wait groups belonging to the connection managers.
Update all public functions that perform ops to first call the
contoller helper to get a provider.
Add wait groups to connection managers and block on them on close.

Change-Id: I17d16de1fa6aebfad09b267aaaebe2424be73893
Reviewed-on: https://review.couchbase.org/c/gocb/+/210690
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>