GOCBC-1664: Fix intermittent expiry assertion failures in range scan tests
Motivation
==========
The clocks on cbdyncluster and the Jenkins hosts can sometimes be out of sync. This results in the expiry assertiosn of TestRangeScanSampling and TestRangeScanRangeWithContentBinaryTranscoder to fail as the absolute expiry coming from the server doesn't match our expectations.

Change
======
Just check that the expiry in the scan results is positive instead of checking that it's at least 15 seconds, which is what we already do for TestRangeScanRangeWithContent.

Results
=======
Tests pass

Change-Id: I8d711c175ed5d125814113df88ae13c28994ed35
Reviewed-on: https://review.couchbase.org/c/gocb/+/215904
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mateusz <matt.wozakowski@couchbase.com>