GOCBC-1662: Allow retries in span assertions for first op in TestAnalyticsIndexesCrud
Motivation
==========
When running tests on a newly created cluster, it's possible that the analytics service is in a 'temporarily unavailable' state briefly. The SDK correctly retries any 'temporary error' analytics status codes, but the assertion for numDispatchSpans can fail because of the retries.

Change
======
* Allow retries in the tracing assertions for the first operation in TestAnalyticsIndexesCrud
* Fix AssertHTTPDispatchSpansGE - assertion was checking that len(spans) <= N, instead of len(spans) >= N

Result
======
Test passes - tested by kicking off the test with a cluster that's undergoing a rebalance

Change-Id: I8dc25bb47a7456e295d27bfa3ecc0af2050c6b27
Reviewed-on: https://review.couchbase.org/c/gocb/+/216095
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>