GOCBC-1667: Close results channel in zone-aware replica reads when server group is not found
Motivation
==========
The logic that checks if the result channel should be closed is being executed when each replica read finishes. This means that if there are no servers to read from (e.g. when the server group does not exist), the channel is never closed and res.Next() in LookupInAllReplicas/GetAllReplicas never returns.

Change
======
In LookupInAllReplicas/GetAllReplicas, close the channels in the result before returning it, if no replica reads are being performed.

Results
=======
Tests pass, including all zone-aware replica read FIT tests

Change-Id: I56807fba7739eeb9784d05613055b7d2bddfa9db
Reviewed-on: https://review.couchbase.org/c/gocb/+/216670
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>