GOCBC-1648: Don't send highlight.fields if not set
Motivation
-----------
When we create the search request payload we are always sending
highlight.fields if highlight is set by the user. If the user hasn't
specified any fields then we send it as null. In order to future
proof a case where the server decides to do something differently
when fields aren't set at all we should only set it in the payload
if the user specifies fields.

Changes
--------
Don't send highlight.fields in the request unless the user specifies
fields.

Change-Id: Ieb6d783d7d06f3aec3d9e6ea9b110f51c5637032
Reviewed-on: https://review.couchbase.org/c/gocb/+/224423
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>