GOCBC-1741: Fix encoding of content in txn insert/replace in query mode
Motivation
==========
With the binary txns changes, a bug was introduced where the content is inserted as a base64 encoded string, instead of the encoded JSON.

Change
======
Cast valueBytes to json.RawMessage before setting it as a positional parameter. Thins ensures that it is set into the request body as is, instead of encoding it as a base64 string.

Results
=======
This resolves all the relevant FIT failures where the content could not be parsed as JSON after inserts/replaces in query mode.

Change-Id: Iea108526f5a93784434ee210987d9675b59ec9b5
Reviewed-on: https://review.couchbase.org/c/gocb/+/229186
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Build Bot <build@couchbase.com>