GOCBC-1758: Update TermRangeQuery to not have a term field
Motivation
----------
We always include a term field in the JSON payload for term range
queries. This is incorrect and leads the fts engine to interpret
the query as a term query. The NewTermRangeQuery signature is
also incorrect, as it should not accept a term parameter.

Changes
-------
Remove the term field from TermRangeQuery.
Update NewTermRangeQuery to not accept any paramters, this is a
breaking change but if users are using this query type then they
are getting incorrect behaviour.

Change-Id: Ib7ac9344b9ab298c499e549e5d240b4056bada13
Reviewed-on: https://review.couchbase.org/c/gocb/+/232151
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Build Bot <build@couchbase.com>