GOCBC-1774: Update full doc projection parsing
Motivation
----------
When we perform a projection with > 16 paths then we do a full
doc fetch and rebuild the value based on what the user requested.
At present we do not correctly build nested values.

Changes
-------
Update full doc projection parsing to recurse the field tree and
use the actual value.

Change-Id: I8ebd70a0f3302f1756480dbbde5a65b167604e01
Reviewed-on: https://review.couchbase.org/c/gocb/+/238786
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>